Make GInetAddress and GSocketAddress not be GInitiallyUnknowned
floating refs only make sense for things that have exactly one owner,
and this is not necessarily true for these types.

In particular, to be able to work with GInetAddresses, GResolver needs
to create temporary GInetSocketAddress objects internally. But
creating a GInetSocketAddress from a GInetAddress will cause the
address to lose its floating ref, which doesn't seem like something
that simply resolving it should do.